requirements.txt
# requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
PyMuPDF==1.23.8
docx2txt==0.8
spacy==3.7.2
joblib==1.3.2
python-dotenv==1.0.0